git-svn-id: https://nocs.stanford.edu/cgi-bin/svn.cgi/booksim2.0/branches/classes@3039 ce7aba0d-9f81-40a6-a495-cecae84b0515

+++ src/injection.cpp
@@ -72,7 +72,7 @@ bool on_off( int source, double rate )
   assert( ( source >= 0 ) && ( source < gNodes ) );
   assert( rate <= 1.0 );
 
-  if ( gNodeStates.size() != gNodes ) {
+  if ( gNodeStates.size() != (size_t)gNodes ) {
     gNodeStates.resize(gNodes, 0);
   }
